% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{add_curvature}
\alias{add_curvature}
\title{Add curvature}
\usage{
add_curvature(d, xvar, yvar)
}
\arguments{
\item{d}{a tibble with two columns, xvar and yvar}

\item{xvar}{a chr column name}

\item{yvar}{a chr column name}
}
\value{
a tibble with a new column, curva
}
\description{
Add curvature
}
\keyword{internal}
