% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{add_dist_2_line}
\alias{add_dist_2_line}
\title{Add distance to line at each x-axis point}
\usage{
add_dist_2_line(fitd, xvar, yvar, p1, p2)
}
\arguments{
\item{fitd}{a tibble from the add_fitted_y}

\item{xvar}{a chr column name}

\item{yvar}{a chr column name}

\item{p1}{NULL, or an int to manually set the first index of line}

\item{p2}{NULL, or an int to manually set the last index of line}
}
\value{
a original fitd + 6 new columns
}
\description{
Add distance to line at each x-axis point
}
\keyword{internal}
