% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_plot.R
\name{add_facet_endpoint_col}
\alias{add_facet_endpoint_col}
\title{Add the facet_endpoint column}
\usage{
add_facet_endpoint_col(statsd)
}
\arguments{
\item{statsd}{the stats component from the list of estimate_dataset_bmr()}
}
\value{
a tibble with some columns are united
}
\description{
For the parameter columns (not including TRSH) and the endpoint column,
if the parameter has more than 1 unique values, this parameter column will be
}
\keyword{internal}
