% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{add_fitted_y}
\alias{add_fitted_y}
\title{Add fitted values by either exponential fit or linear fit to the original table}
\usage{
add_fitted_y(d, xvar, yvar)
}
\arguments{
\item{d}{a tibble with two columns, xvar and yvar}

\item{xvar}{a chr column name}

\item{yvar}{a chr column name}
}
\value{
the original d + y_exp_fit + y_lm_fit columns
}
\description{
Add fitted values by either exponential fit or linear fit to the original table
}
\keyword{internal}
