% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{add_hit}
\alias{add_hit}
\title{A wrapper function for the add_hit_actset}
\usage{
add_hit(nestd)
}
\arguments{
\item{nestd}{the nested tibble from the get_nested_joined_sets}
}
\value{
the nested tibble but in each table of the act_set, a new column hit is added
}
\description{
A wrapper function for the add_hit_actset
}
\keyword{internal}
