% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{add_hit_actset}
\alias{add_hit_actset}
\title{Add hit column for the act_set}
\usage{
add_hit_actset(act_set)
}
\arguments{
\item{act_set}{one act_set}
}
\value{
a tibble of the act_set with a new column hit; 1 as active
}
\description{
wAUC != 0 as active otherwise inactive
}
\keyword{internal}
