% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{apply_comment_to_unhit}
\alias{apply_comment_to_unhit}
\title{Use the Comment and hit column to adjust the activity call}
\usage{
apply_comment_to_unhit(nestd, act_modifier)
}
\arguments{
\item{nestd}{the nested tibble from the get_nested_joined_sets and the act_set has the hit column}

\item{act_modifier}{default = NULL, no modification on the activity}
}
\value{
the same tibble but column values could be modified
}
\description{
Use the Comment and hit column to adjust the activity call
}
\keyword{internal}
