% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{apply_comment_to_unhit_actset}
\alias{apply_comment_to_unhit_actset}
\title{Use the comment to unhit values}
\usage{
apply_comment_to_unhit_actset(act_set, act_modifier)
}
\arguments{
\item{act_set}{one act_set}

\item{act_modifier}{a character string}
}
\value{
a tibble with all activity related columns modified
}
\description{
Use the comment to unhit values
}
\keyword{internal}
