% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{apply_comment_to_unhit_nonactset}
\alias{apply_comment_to_unhit_nonactset}
\title{Use the hit column from act_set to adjust the corrected_resp in resp_set or Cxx/ECxx in fp_set}
\usage{
apply_comment_to_unhit_nonactset(nonact_set, act_set)
}
\arguments{
\item{act_set}{one act_set}

\item{resp_set}{either one resp_set or one fp_set}
}
\value{
the modified nonactset
}
\description{
Use the hit column from act_set to adjust the corrected_resp in resp_set or Cxx/ECxx in fp_set
}
\keyword{internal}
