% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_simulate.R
\name{assign_dataset_type}
\alias{assign_dataset_type}
\title{Assign dataset type
(use the column names of a dataset to determine the type)}
\usage{
assign_dataset_type(d)
}
\arguments{
\item{d}{a dataset such as \code{\link{zfishdev}} and \code{\link{zfishbeh}}}
}
\value{
a character (dichotomous or continuous)
}
\description{
Assign dataset type
(use the column names of a dataset to determine the type)
}
\keyword{internal}
