% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{cal_curvature}
\alias{cal_curvature}
\title{Calculate curvature}
\usage{
cal_curvature(thres, vars)
}
\arguments{
\item{thres}{x-axis}

\item{vars}{y-axis}
}
\value{
a vector of curvature at each x point
}
\description{
https://en.wikipedia.org/wiki/Curvature
}
\keyword{internal}
