% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_onerun.R
\name{cal_curvep_dataset}
\alias{cal_curvep_dataset}
\title{Prepare the nested structure of for curvep and calculate (call_curvep)}
\usage{
cal_curvep_dataset(d, config)
}
\arguments{
\item{d}{a dataset with columns: endpoint, chemical, conc, and resp, mask}

\item{config}{curvep_defaults()}
}
\value{
a tibble with a new column output
}
\description{
Prepare the nested structure of for curvep and calculate (call_curvep)
}
\keyword{internal}
