% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{cal_dataset_knee}
\alias{cal_dataset_knee}
\title{Calculate knee point for each dataset}
\usage{
cal_dataset_knee(pvard, base_cols, p1, p2)
}
\arguments{
\item{pvard}{output cal_dataset_pvar()}

\item{base_cols}{parameter columns + endpoint + chemical, no sample_id}

\item{p1}{NULL, or an int to manually set the first index of line}

\item{p2}{NULL, or an int to manually set the last index of line}
}
\value{
the pvard + a new column knee_out
}
\description{
Calculate knee point for each dataset
}
\keyword{internal}
