% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{cal_dataset_pvar}
\alias{cal_dataset_pvar}
\title{Calculate pooled variance column}
\usage{
cal_dataset_pvar(act_set, base_cols)
}
\arguments{
\item{act_set}{summarize_rcurvep_output()$result$act_set}

\item{base_cols}{parameter columns + endpoint + chemical, no sample_id}
}
\value{
a tibble with parameter columns + endpoint + pvar
}
\description{
Calculate pooled variance column
}
\keyword{internal}
