% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{cal_dist2l}
\alias{cal_dist2l}
\title{Calculate the distance to a line}
\usage{
cal_dist2l(thres, vars, p1 = NULL, p2 = NULL)
}
\arguments{
\item{thres}{x-axis}

\item{vars}{y-axis}

\item{p1}{NULL, or an int to manually set the first index of line}

\item{p2}{NULL, or an int to manually set the last index of line}
}
\value{
a vector of curvature at each x point
}
\description{
https://dataplatform.cloud.ibm.com/analytics/notebooks/54d79c2a-f155-40ec-93ec-ed05b58afa39/view?access_token=6d8ec910cf2a1b3901c721fcb94638563cd646fe14400fecbb76cea6aaae2fb1
}
\keyword{internal}
