% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{cal_exponential_fit}
\alias{cal_exponential_fit}
\title{Calculate exponential fit (y ~ x)}
\usage{
cal_exponential_fit(thres, vars)
}
\arguments{
\item{thres}{x-axis}

\item{vars}{y-axis}
}
\value{
an nls object
}
\description{
http://douglas-watson.github.io/post/2018-09_exponential_curve_fitting/&sa=D&source=hangouts&ust=1547125299019000&usg=AFQjCNGzoP26jIUYQnAS9Z5Eb4aO66IzPg
}
\keyword{internal}
