% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR_calculations.R
\name{cal_exponential_inflection}
\alias{cal_exponential_inflection}
\title{Calculate the inflection point on the exponential curve using approaches
(currently only distance-based approach is implemented)}
\usage{
cal_exponential_inflection(df, xvar, yvar, p1 = NULL, p2 = NULL)
}
\arguments{
\item{df}{a tibble}

\item{xvar}{a character column name in the `df` to be the x-axis in the exponential curve}

\item{yvar}{a character column name in the `df` to be the y-axis in the expoential curve}

\item{p1}{default = NULL, or a number to manually set the first threshold index for the distance approach}

\item{p2}{default = NULL, or a number to manually set last threshold index for the distance approach}
}
\value{
a tibble with added columns
\itemize{
  \item thresDist: the identified threshold based on distance approach
  \item thresDistComment: the flag to suggest whether to use the thresDist (OK, cautionary, inadvisable)
  \item p1, p2, distl: the index of points for the line and the calculated distance-to-line value
  \item curvature: the calculated curvature value
}
}
\description{
Calculate the inflection point on the exponential curve using approaches
(currently only distance-based approach is implemented)
}
\examples{
data("zfishdev_act")
acts <- zfishdev_act \%>\% mutate(POD = ifelse(is.na(POD), conc_highest, POD))
outthres <- identify_basenoise_threshold(acts)

outthres2 <- outthres \%>\%
split(., list(.$endpoint, .$direction), drop = TRUE) \%>\%
map_df(cal_exponential_inflection, xvar = "threshold", yvar = "pooled_variance")

}
