% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{cal_linear_fit}
\alias{cal_linear_fit}
\title{Calcualte linear fit (y ~ x)}
\usage{
cal_linear_fit(thres, vars)
}
\arguments{
\item{thres}{x-axis}

\item{vars}{y-axis}
}
\value{
an lm object
}
\description{
Calcualte linear fit (y ~ x)
}
\keyword{internal}
