% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_simulate.R
\name{cal_lm_pred_resp}
\alias{cal_lm_pred_resp}
\title{Calculate predicted resp from the linear regressinon fitting (conc x resp)}
\usage{
cal_lm_pred_resp(d)
}
\arguments{
\item{d}{a dataset such as \code{\link{zfishdev}} and \code{\link{zfishbeh}}}
}
\value{
d with new resp column
}
\description{
Calculate predicted resp from the linear regressinon fitting (conc x resp)
}
\keyword{internal}
