% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_simulate.R
\name{cal_percent_resps}
\alias{cal_percent_resps}
\title{Calculate percent of response by bootstrap}
\usage{
cal_percent_resps(n_in, N, times)
}
\arguments{
\item{n_in}{number of incidence}

\item{N}{total number of animals}

\item{times}{bootstrap times}
}
\value{
a vector of bootstrap percent of response
}
\description{
Calculate percent of response by bootstrap
}
\keyword{internal}
