% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_onerun.R
\name{call_curvep}
\alias{call_curvep}
\title{Call curvep}
\usage{
call_curvep(concs, resps, masks = NULL, paras = curvep_defaults())
}
\arguments{
\item{concs}{a vector of concs (from lowest to highest concentration)}

\item{resps}{a vector of resps (from lowest to highest concentration)}

\item{masks}{a vector of masks}

\item{paras}{curvep_defaults() or a modified list}
}
\value{
(see \code{\link{curvep}}) but remove "Settings"
}
\description{
Call curvep
}
\keyword{internal}
