% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_simulate.R
\name{clean_add_sampleid}
\alias{clean_add_sampleid}
\title{Clean and add sample_id column}
\usage{
clean_add_sampleid(d, n_samples)
}
\arguments{
\item{d}{a tibble}

\item{n_samples}{number of samples}
}
\value{
d with sample_id
}
\description{
Clean and add sample_id column
}
\keyword{internal}
