% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_onerun.R
\name{clean_curvep_output}
\alias{clean_curvep_output}
\title{Clean Curvep output to extract the useful information}
\usage{
clean_curvep_output(d, config)
}
\arguments{
\item{d}{a tibble from cal_curvep_dataset()}

\item{config}{curvep_defaults()}
}
\value{
a tibble with new columns, out_resp, in_summary, fingerprint, activity
}
\description{
Clean Curvep output to extract the useful information
}
\keyword{internal}
