% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_combirun.R
\name{combi_run_curvep_in}
\alias{combi_run_curvep_in}
\title{Run curvep using created curvep parameters on}
\usage{
combi_run_curvep_in(d, mask, n_samples, keep_sets, paras)
}
\arguments{
\item{d}{a dataset after \code{\link{create_datasets}}}

\item{mask}{0 (default) for not using mask}

\item{n_samples}{NULL (default) or an int to indicate the number of resp per conc to simulate}
}
\value{
a tibble (paras) with a new rcurvep_obj column
}
\description{
Run curvep using created curvep parameters on
}
\keyword{internal}
