% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_combirun.R
\name{combi_run_rcurvep}
\alias{combi_run_rcurvep}
\title{Run rcurvep on a base/simulated dataset with a combination of parameters}
\usage{
combi_run_rcurvep(d, n_samples = NULL, vdata = NULL, mask = 0,
  keep_sets = c("act_set", "resp_set", "fp_set"), ...)
}
\arguments{
\item{d}{a dataset such as \code{\link{zfishdev}} and \code{\link{zfishbeh}}}

\item{n_samples}{NULL (default) or an int to indicate the number of resp per conc to simulate}

\item{vdata}{NULL (default) or a dbl vector of responses in vehicle control wells; experimental feature}

\item{mask}{0 (default) for not using mask
use a vector of integers to mask the resps, 1 to mask resp at the highest conc, 2 to mask resp at the second highest conc, and so on.}

\item{keep_sets}{act_set + (resp_set or fp_set) or all three
 \itemize{
  \item act_set: activity data
  \item resp_set: response data
  \item fp_set: fingerprint data
}}

\item{...}{a list of settings; are used to overwrite the default values in curvep_defaults()}
}
\value{
a list with two components: result (a named list, based on keep_data), config
}
\description{
A wrapper function combining \code{\link{create_dataset}}, \code{\link{run_rcurvep}}
}
\examples{

data(zfishbeh)

out <- combi_run_rcurvep(zfishbeh, n_samples = 2, TRSH = c(5, 10), mask = 1, keep_sets = "act_set")



}
