% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{create_bmr_report}
\alias{create_bmr_report}
\title{Create the output for BMR from two approaches}
\usage{
create_bmr_report(distd, xvar, yvar)
}
\arguments{
\item{distd}{the output of add_dist_2_line()}

\item{xvar}{a chr column name}

\item{yvar}{a chr column name}
}
\value{
a tibble with generate columns
}
\description{
Create the output for BMR from two approaches
}
\keyword{internal}
