% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_plot.R
\name{create_bmrplot_based}
\alias{create_bmrplot_based}
\title{Create the basic tibble for the BMR diagnostic plot}
\usage{
create_bmrplot_based(statsd)
}
\arguments{
\item{statsd}{the output from add_facet_endpoint_col()}
}
\value{
a tibble facet_endpoint, TRSH, pvar, dist2l_ori, dist2l_exp, y_exp_fit
}
\description{
Create the basic tibble for the BMR diagnostic plot
}
\keyword{internal}
