% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_simulate.R
\name{create_dataset}
\alias{create_dataset}
\title{Create a dataset for rcurvep either by summarizing the resp data or by simulating the resp data}
\usage{
create_dataset(d, n_samples = NULL, vdata = NULL)
}
\arguments{
\item{d}{a dataset such as \code{\link{zfishdev}} and \code{\link{zfishbeh}}}

\item{n_samples}{NULL (default) or an int to indicate the number of resp per conc to simulate}

\item{vdata}{NULL (default) or a dbl vector of responses in vehicle control wells;
this parameter only works when n_samples is not NULL; an experimental feature}
}
\value{
d with sample_id (if n_samples is not NULL) or d
}
\description{
Summary: (n_samples = NULL)
A base dataset: only one response per endpoint-chemical-concentration
For dichotomous data, percentage is reported (n_in/N*100)
For continuous data, median resp per conc is reported
}
\details{
Simulation: (n_samples is a positive integer)
A simulated dataste: allow multiple responses per endpoint-chemical-concentration
For dichotomous data, bootstrap is used on the n_in vector to create a vector of percent response
FOr continuous data, options a) direct sampling; b) linear fit + error from vehicle control data
}
\examples{

# continous
data(zfishbeh)

## default
d <- create_dataset(zfishbeh)

## add samples
d <- create_dataset(zfishbeh, n_samples = 3)

## add samples and vdata
d <- create_dataset(zfishbeh, n_samples = 3, vdata = rnorm(100))

# dichotomous
data(zfishdev)

## default
d <- create_dataset(zfishdev)

## add samples
d <- create_dataset(zfishdev, n_samples = 3)

}
