% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplfit_onerun.R
\name{create_hillsimu_resp}
\alias{create_hillsimu_resp}
\title{Create simulated responses by the hill fit results.}
\usage{
create_hillsimu_resp(inp, out)
}
\arguments{
\item{inp}{A list (tibble) of with conc and resp columns.}

\item{out}{A list of named models.}
}
\value{
The vector of responses.
}
\description{
The new responses is the fitted responses + sampled residuals.
}
\keyword{internal}
