% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_simulate.R
\name{create_n_errors}
\alias{create_n_errors}
\title{create_n_errors by sample with replacement}
\usage{
create_n_errors(vec, n = 100)
}
\arguments{
\item{vec}{a vector of responses}

\item{n}{number of samples}
}
\value{
a vector
}
\description{
create_n_errors by sample with replacement
}
\keyword{internal}
