% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_combirun.R
\name{create_para_input}
\alias{create_para_input}
\title{create parameter input}
\usage{
create_para_input(paras, n_samples, d)
}
\arguments{
\item{paras}{curvep parameters}

\item{n_samples}{NULL or an int to indicate the number of resp per conc to simulate}

\item{d}{a dataset with columns: endpoint, chemical, conc, and resp, mask (optional) (see \code{\link{zfishbeh}})}
}
\value{
a tibble with all combinations of parameters in a data column + sample_id (for a simulated dataset)
or a tibble with all combinations of parameters
}
\description{
create parameter input
}
\keyword{internal}
