% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_onerun.R
\name{create_resp_mask}
\alias{create_resp_mask}
\title{Creat response mask}
\usage{
create_resp_mask(d, mask)
}
\arguments{
\item{d}{a dataset with columns: endpoint, chemical, conc, and resp, mask (optional) (see \code{\link{zfishbeh}})}

\item{mask}{0 for no mask; NULL no change of the data
use a vector of integers to mask the resps, 1 to mask resp at the highest conc, 2 to mask resp at the second highest conc, and so on.}
}
\value{
a tibble with added mask column
}
\description{
Creat response mask
}
\keyword{internal}
