% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_simulate.R
\name{create_resps}
\alias{create_resps}
\title{A wrap function to determine whether to use summarize_resps or simulate_resps}
\usage{
create_resps(d, dataset_type, n_samples, vdata)
}
\arguments{
\item{d}{a dataset such as \code{\link{zfishdev}} and \code{\link{zfishbeh}}}

\item{dataset_type}{a character (dichotomous or continuous)}

\item{n_samples}{NULL (default) or an int to indicate the number of resp per conc to simulate}

\item{vdata}{NULL (default) or a dbl vector of responses in vehicle control wells}
}
\value{
d with a modified or created resp column and/or sample_id
}
\description{
A wrap function to determine whether to use summarize_resps or simulate_resps
}
\keyword{internal}
