% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_simulate.R
\name{define_quantile_outliers}
\alias{define_quantile_outliers}
\title{Define outliers by quantile and a set cutoff}
\usage{
define_quantile_outliers(vec, degree = 3, invalid_thr = 0.05)
}
\arguments{
\item{vec}{a vector of responses}

\item{degree}{x IQR (default = 3)}

\item{invalid_thr}{(default = 0.05)}
}
\value{
a logical vec with TRUE as the outliers
}
\description{
if x fraction of outliers exist, the Tukey approach is not suitable.
}
\keyword{internal}
