% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{estimate_dataset_bmr}
\alias{estimate_dataset_bmr}
\title{Estimate BMR for each dataset}
\usage{
estimate_dataset_bmr(d, p1 = NULL, p2 = NULL, plot = TRUE)
}
\arguments{
\item{d}{the output from \code{\link{summarize_rcurvep_output}}}

\item{p1}{default = NULL, or an int to manually set the first index of line}

\item{p2}{default = NULL, or an int to manually set the last index of line}

\item{plot}{default = TRUE, plot the diagnostic plot}
}
\value{
a list with two components: stats and outcome
}
\description{
Estimate BMR for each dataset
}
\examples{

data(zfishdev_act)

# use the highest concentration to replace value of the inactive
sumd <- summarize_rcurvep_output(zfishdev_act, clean_only = TRUE)

bmr_out <- estimate_dataset_bmr(sumd, plot = FALSE)
plot(bmr_out)

}
