% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_onerun.R
\name{extract_curvep_activity}
\alias{extract_curvep_activity}
\title{Extract Curvep activity}
\usage{
extract_curvep_activity(x, config)
}
\arguments{
\item{x}{the list from curvep output}

\item{config}{for the DUMV parameter}
}
\value{
a tibble with activity
}
\description{
Extract Curvep activity
}
\keyword{internal}
