% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvep_out_extraction.R
\name{extract_curvep_data}
\alias{extract_curvep_data}
\title{Unnest Curvep input, output, or activity lists}
\usage{
extract_curvep_data(c_out, type)
}
\arguments{
\item{c_out}{A tibble after `run_curvep_job()`.}

\item{type}{A string to indicate which data to extract. Currently six types are implemented:
act: all the activity related metrics such as potency and efficacy
concs_hl: the highest and lowest tested concentration
concs_in: a vector of input concentrations
resps_out: a vector of output (clean) responses
paras_in: manually input parameters
paras: all the parameters used in the calculation}
}
\value{
Depending the specified type, a tibble with various columns is returned.
}
\description{
The function unnest the input, output, or activity lists after `run_curvep_job()`
}
\examples{
data(zfishdev)
}
\seealso{
\code{\link{curvep}} for available Curvep parameters
}
