% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_onerun.R
\name{extract_curvep_fingerprint}
\alias{extract_curvep_fingerprint}
\title{Extract Curvep finterprint}
\usage{
extract_curvep_fingerprint(x, config)
}
\arguments{
\item{x}{the list from curvep output}

\item{config}{to get the DUMV parameter}
}
\value{
a tibble with three columns: xx, ECxx, Cxx
}
\description{
Extract Curvep finterprint
}
\keyword{internal}
