% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_onerun.R
\name{extract_curvep_outresp}
\alias{extract_curvep_outresp}
\title{Extract Curvep response output}
\usage{
extract_curvep_outresp(x)
}
\arguments{
\item{x}{the list from curvep output}
}
\value{
a tibble with two columns: corrected_resp (resp) and correction (corr)
}
\description{
Extract Curvep response output
}
\keyword{internal}
