% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_onerun.R
\name{extract_input_summary}
\alias{extract_input_summary}
\title{Extract input (conc, resp) summary data}
\usage{
extract_input_summary(x)
}
\arguments{
\item{x}{the tibble of the input}
}
\value{
a tibble
}
\description{
Extract input (conc, resp) summary data
}
\keyword{internal}
