% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_combirun.R
\name{flat_result_tbl}
\alias{flat_result_tbl}
\title{Map through the rcurvep objects in the tibble and unnest}
\usage{
flat_result_tbl(d, keep_set)
}
\arguments{
\item{d}{a tibble (paras) with a new rcurvep_obj column \code{\link{combi_run_curvep_in}}}

\item{keep_set}{only one is allowed}
}
\value{
a tibble
}
\description{
Map through the rcurvep objects in the tibble and unnest
}
\keyword{internal}
