% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR_calculations.R
\name{generate_diagnostic_plot}
\alias{generate_diagnostic_plot}
\title{Generate diagnostic plots for inspecting the quality of the identified baseline noise threshold}
\usage{
generate_diagnostic_plot(df, endpoint = "endpoint", direction = "direction",
  threshold = "threshold", n_endpoint_page = 4)
}
\arguments{
\item{df}{a tibble from the identify_basenoise_threshold() or a tibble with two required columns: pooled_variance and dist2l}

\item{endpoint}{a string that represents the column name of endpoint in `df`}

\item{direction}{a string that represents the column name of direction in `df`}

\item{threshold}{a string that represents the column name of threshold in `df`}

\item{n_endpoint_page}{number of endpoints to be plotted per page}
}
\value{
a ggplot object
}
\description{
Generate diagnostic plots for inspecting the quality of the identified baseline noise threshold
}
\examples{
data("zfishdev_act")
acts <- zfishdev_act \%>\% mutate(POD = ifelse(is.na(POD), conc_highest, POD))
outthres <- identify_basenoise_threshold(acts, plot = FALSE)
generate_diagnostic_plot(outthres)

}
