% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{get_base_cols}
\alias{get_base_cols}
\title{Get common column names between sets}
\usage{
get_base_cols(lsets)
}
\arguments{
\item{lsets}{The result list from the combi_run_rcurvep or run_rcurvep}
}
\value{
a vector of common column names in sets
}
\description{
Depending on different settings (e.g., TRSH), the common column names change
}
\keyword{internal}
