% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR_calculations.R
\name{get_baseline_threshold}
\alias{get_baseline_threshold}
\title{Get baseline noise threshold through activity data calculated from simulated curves}
\usage{
get_baseline_threshold(df, endpoint = "endpoint", chemical = "chemical",
  threshold = "threshold", direction = "direction", potency = "POD",
  plot = TRUE, n_endpoint_page = 4)
}
\arguments{
\item{df}{a tbl}

\item{endpoint}{a chr that represents the column name of endpoint in `df`}

\item{chemical}{a chr that represents the column name of chemical in `df`}

\item{threshold}{a chr that represents the column name of threshold in `df`}

\item{direction}{a chr that represents the column name of direction in `df`}

\item{potency}{a chr that represents the column name of potency in `df`}

\item{plot}{default = TRUE, for diagnostic plots}

\item{n_endpoint_page}{number of endpoints to be plotted per page}
}
\value{
an object of class 'rcurvep_thres_stats' with two named tbls:

\describe{
  \item{stats}{
     \itemize{
       \item y_exp_fit, y_lm_fit: the y value from the fitting results (exponential or linear)
       \item dist2l_exp, dist2l_raw: the calculated distance-to-line value
       based on the y from exponential fit or original y
       \item curvature: the calculated curvature value using smooth.spline fit results
       \item p1_raw, p2_raw: the index of points for the line using the original y values
     }
  }
  \item{outcome}{
    \itemize{
       \item thresDist_raw: the identified threshold based on distance approach using the original y values
       \item thresDist_exp: the identified threshold based on distance approach using the y values from the exponential fitting results
       \item thresComment: the flag to suggest whether to use the thresDist (OK, cautionary, check)
       \item p1_raw, p2_raw: the index of points for the line using the original y values
       \item cor_exp_fit, cor_lm_fit: the Pearson's correlation between the y values from fitting (exponential or linear) and the original y values
    }
  }
}
}
\description{
Given an activity dataset after `run_curvep_batch()` and `withdraw()`,
the function calculates the pooled variance of potency information (i.e., POD parameter) across chemicals
derives the baseline noise threshold, which is the lowest threshold that variance of POD is sufficiently reduced and stabilized.
}
\examples{
data("zfishdev_act")
outthres <- get_baseline_threshold(zfishdev_act)

}
