% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_plot.R
\name{get_dataset_linecoeff}
\alias{get_dataset_linecoeff}
\title{Get line coefficients}
\usage{
get_dataset_linecoeff(statsd, fit_type = c("ori", "exp"))
}
\arguments{
\item{statsd}{the output from add_facet_endpoint_col()}

\item{fit_type}{either ori or exp}
}
\value{
a tibble with type, fit_type, intercept, slope
}
\description{
Get line coefficients
}
\keyword{internal}
