% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_plot.R
\name{get_facet_endpoints_in_page}
\alias{get_facet_endpoints_in_page}
\title{Split the facet_endpoints into pages}
\usage{
get_facet_endpoints_in_page(plotd, n_in_page)
}
\arguments{
\item{plotd}{output from create_bmrplot_based}

\item{n_in_page}{number of datasets in a page}
}
\value{
a list, each component has the facet_endpoints in a page
}
\description{
Split the facet_endpoints into pages
}
\keyword{internal}
