% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{get_nested_joined_sets}
\alias{get_nested_joined_sets}
\title{Nest the lsets into columns}
\usage{
get_nested_joined_sets(lsets, base_cols)
}
\arguments{
\item{lsets}{The result list from the combi_run_rcurvep or run_rcurvep}

\item{base_cols}{a vector of common column names in sets}
}
\value{
a tibble with the common column from sets + the data of sets nested into columns
}
\description{
Nest the lsets into columns
}
\keyword{internal}
