% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_plot.R
\name{get_p1_p2_linecoeff}
\alias{get_p1_p2_linecoeff}
\title{Calculate the intercept and slope for the line by p1 and p2}
\usage{
get_p1_p2_linecoeff(dd, xvar, yvar, p1, p2)
}
\arguments{
\item{dd}{dataset}

\item{xvar}{x-axis}

\item{yvar}{y-axis}

\item{p1}{an int for the first index of line}

\item{p2}{an int for the last index of line}
}
\value{
a data frame with intercept and slope columns
}
\description{
Calculate the intercept and slope for the line by p1 and p2
}
\keyword{internal}
