% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{get_thres_at_max_dist2l}
\alias{get_thres_at_max_dist2l}
\title{Get the threshold at which has the max distance to the line}
\usage{
get_thres_at_max_dist2l(thres, dist2l)
}
\arguments{
\item{thres}{a vector of thresholds}

\item{dist2l}{a vector of distances}
}
\value{
a single value from the thres
}
\description{
Get the threshold at which has the max distance to the line
}
\keyword{internal}
