% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR_calculations.R
\name{identify_basenoise_threshold}
\alias{identify_basenoise_threshold}
\title{Identify baseline noise threshold through activity data from simulated curves}
\usage{
identify_basenoise_threshold(df, endpoint, chemical, threshold, direction,
  potency, plot = TRUE, n_endpoint_page = 4)
}
\arguments{
\item{df}{a tibble}

\item{endpoint}{a string that represents the column name of endpoint in `df`}

\item{chemical}{a string that represents the column name of chemical in `df`}

\item{threshold}{a string that represents the column name of threshold in `df`}

\item{direction}{a string that represents the column name of direction in `df`}

\item{potency}{a string that represents the column name of potency in `df`}

\item{plot}{default = TRUE, diagnostic plot}

\item{n_endpoint_page}{number of endpoints to be plotted per page}
}
\value{
a tibble with added columns
\itemize{
  \item pooled_variance: the pooled varience of the potency
  \item thresDist: the identified threshold based on distance approach
  \item thresCurva: the identified threshold based on curvature + distance approach
  \item p1, p2, distl: the index of points for the line and the calculated distance-to-line value
  \item curvature: the calculated curvature value
}
}
\description{
Provided an activity dataset after `run_curvep_job()` and `extract_curvep_data()`,
the function calculates the pooled variance of potency information (i.e., POD parameter) across chemicals
and approaches are used to derive the lowest threshold that variance of POD is sufficiently reduced and even stabilized.
}
\examples{
vignette("Rcurvep-intro")

}
