% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{make_act_na_highconc}
\alias{make_act_na_highconc}
\title{Replace all NA in the activity related columns as highest tested concentration (wrapper)}
\usage{
make_act_na_highconc(nestd)
}
\arguments{
\item{nestd}{the nested tibble from the get_nested_joined_sets}
}
\value{
the same nested tibble but act_set has been modified
}
\description{
Replace all NA in the activity related columns as highest tested concentration (wrapper)
}
\keyword{internal}
