% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{make_act_na_highconc_in}
\alias{make_act_na_highconc_in}
\title{Replace all NA in the activity related columns as highest tested concentration}
\usage{
make_act_na_highconc_in(act_set)
}
\arguments{
\item{act_set}{one act_set}
}
\value{
an act_set with modified columns (wConc, EC50, C50, POD)
}
\description{
Replace all NA in the activity related columns as highest tested concentration
}
\keyword{internal}
