% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_combirun.R
\name{merge_rcurvep_output}
\alias{merge_rcurvep_output}
\title{Merge columns from the result field of an rcurvep object}
\usage{
merge_rcurvep_output(d, keep_sets)
}
\arguments{
\item{keep_sets}{one or multiple of these: act_set, resp_set, fp_set
 \itemize{
  \item act_set: activity data
  \item resp_set: response data
  \item fp_set: fingerprint data
}}

\item{rcurvep_obj}{an object with class:rcurvep}
}
\value{
a list with two components: result (a named list, act_set, resp_set, fp_set), config
}
\description{
Merge columns from the result field of an rcurvep object
}
\examples{

data(zfishdev)
d <- create_dataset(zfishdev)
out <- run_rcurvep(d)
res <- merge_rcurvep_output(out)

}
