% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_plot.R
\name{plot.rcurvep_bmr}
\alias{plot.rcurvep_bmr}
\title{Plot BMR diagnostic curves}
\usage{
\method{plot}{rcurvep_bmr}(x, ...)
}
\arguments{
\item{x}{The rcurvep_bmr object from \code{\link[=estimate_dataset_bmr]{estimate_dataset_bmr()}}.}

\item{...}{Allowed values: n_in_page, number of endpoints in a page.}
}
\value{
A ggplot object.
}
\description{
Plot BMR diagnostic curves
}
\examples{

data(zfishdev_act)
bmr_out <- estimate_dataset_bmr(zfishdev_act, plot = FALSE)
plot(bmr_out)

}
