% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR_calculations.R
\name{plot.rcurvep_thres_stats}
\alias{plot.rcurvep_thres_stats}
\title{Plots for inspecting the quality of the derived baseline noise thresholds}
\usage{
\method{plot}{rcurvep_thres_stats}(x, ...)
}
\arguments{
\item{x}{the output from `get_baseline_threshold()` or a tbl similar to the stats output from `get_baseline_threshold()`}

\item{...}{\code{\link{get_baseline_threshold}} for available parameters}
}
\value{
a ggplot object
}
\description{
Plots for inspecting the quality of the derived baseline noise thresholds
}
\examples{
data("zfishdev_act")
outthres <- get_baseline_threshold(zfishdev_act, plot = FALSE)
plot(outthres)

}
