% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_plot.R
\name{plot_diagnostic}
\alias{plot_diagnostic}
\title{Plot the diagnostic plot}
\usage{
plot_diagnostic(plotd, lined, endpoints)
}
\arguments{
\item{lined}{from get_p1_p2_linecoeff()}

\item{endpoints}{the facet_endpoint names in a page}
}
\value{
a ggplot object
}
\description{
Plot the diagnostic plot
}
\keyword{internal}
