% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_combirun.R
\name{pmap_run_rcurvep}
\alias{pmap_run_rcurvep}
\title{Run rcurvep using the parameters supplied by purrr::pmap}
\usage{
pmap_run_rcurvep(d, mask, n_samples, keep_sets, ...)
}
\arguments{
\item{d}{a dataset after \code{\link{create_datasets}}}

\item{mask}{0 (default) for not using mask}

\item{n_samples}{NULL (default) or an int to indicate the number of resp per conc to simulate}
}
\value{
an rcurvep object
}
\description{
Run rcurvep using the parameters supplied by purrr::pmap
}
\keyword{internal}
