% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{qc_data}
\alias{qc_data}
\title{Quality control on the data (pvar ~ TRSH) for getting the BMR}
\usage{
qc_data(exp_cor, linear_cor)
}
\arguments{
\item{exp_cor}{correlation coefficient between pvar and y_exp_fit}

\item{linear_cor}{correlation coefficient between pvar and y_lm_fit}
}
\value{
a character string either of three values: OK, cautionary, check
}
\description{
By using the correlation coefficient between pvar and y_exp_fit and y_lm_fit,
to heuristically QC the data
}
\keyword{internal}
