% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvep_job_wrap_reparam.R
\name{reparam}
\alias{reparam}
\alias{reparam.rcurvep_out_nested}
\title{Re-paramaterization curvep based on output data from `run_curvep_batch()`}
\usage{
reparam(dats_out, directionality = NULL, threshold = NULL,
  other_paras = list(), simplify_output = FALSE)

\method{reparam}{rcurvep_out_nested}(dats_out, directionality = NULL,
  threshold = NULL, other_paras = list(), simplify_output = FALSE)
}
\arguments{
\item{dats_out}{the output from run_curvep_batch(simplify_output = FALSE)}

\item{directionality}{NULL for not changing the directionality, only 1 or -1 is allowed}

\item{threshold}{NULL for not changing the threshold, only one positive dbl threshold is allowed}

\item{other_paras}{a list of other curvep parameters to pass on}

\item{simplify_output}{default = FALSE, if set TRUE (withdraw(out, "act") is performed)}
}
\value{
see \code{\link{run_curvep_batch}}
}
\description{
Given the output from `run_curvep_batch(simplify_output = FALSE)`,
the function re-parameterize input for `curvep()`,
performs calculations, and generates output.
}
\examples{
data(zfishdev)
outd <- run_curvep_batch(zfishdev,
                      directionality = 1,
                      n_sample = 1,
                      threshold = 15,
                      other_paras = list(CARR = 20, TrustHi = TRUE))
outd2 <- reparam(outd, threshold = 25)
# more examples are availabie
vignette("Rcurvep-intro")


}
