% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvep_job_wrap_reparam.R
\name{reparam_curvep_job}
\alias{reparam_curvep_job}
\title{Re-paramaterization based on output data frame from run_curvep_job()}
\usage{
reparam_curvep_job(dats_out, directionality = NULL, threshold = NULL,
  other_paras = list(), simplify_output = FALSE)
}
\arguments{
\item{dats_out}{complex output structure from run_curvep_job(simplify_output = FALSE)}

\item{directionality}{NULL for not changing the directionality, only 1 or -1 is allowed}

\item{threshold}{NULL for not changing the threshold, only one positive numeric threshold is allowed}

\item{other_paras}{a list of other Curvep parameters to pass on}

\item{simplify_output}{(default = FALSE) warning! large output, set TRUE (extract_curvep_data(out, "act") is performed) for threshold finding calculations}
}
\value{
see \code{\link{run_curvep_job}}
}
\description{
Given the complex output structure from run_curvep_job(simplify_output = FALSE),
the function re-parameterize input for `curvep()`,
performs calculations, and generates output.
}
\examples{
data(zfishdev)
outd <- run_curvep_job(zfishdev,
                      directionality = 1,
                      n_sample = 1,
                      threshold = 15,
                      other_paras = list(CARR = 20, TrustHi = TRUE))
outd2 <- reparam_curvep_job(outd, threshold = 25)
# more examples are availabie
vignette("Rcurvep-intro")


}
