% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{replace_active_value}
\alias{replace_active_value}
\title{Replace the original value if there is a string in the comment}
\usage{
replace_active_value(act_modifier, comment, ori_value, new_value)
}
\arguments{
\item{act_modifier}{a character string}

\item{comment}{the comment string}

\item{ori_value}{original values}

\item{new_value}{new value}
}
\value{
a vector with some original values being modified
}
\description{
Replace the original value if there is a string in the comment
}
\keyword{internal}
