% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvep_job_wrap.R
\name{run_curvep_batch}
\alias{run_curvep_batch}
\title{Run curvep in batch mode based on the type of dataset}
\usage{
run_curvep_batch(dats, directionality = c(1, 0, -1), n_sample = NULL,
  threshold, other_paras = list(), vehicle_data = NULL,
  simplify_output = FALSE)
}
\arguments{
\item{dats}{datasets such as \code{\link{zfishdev}} and \code{\link{zfishbeh}}}

\item{directionality}{an int (1, 0, -1) to represent the presumed direction of responses for processing; 1 = up, -1 = down, 0 = both}

\item{n_sample}{NULL (using original data) or an int to indicate the number of curves to simulate}

\item{threshold}{a dbl for the presumed noise threshold or a dbl vector for threshold finding or a named (-1, 1) list with dbl vectors}

\item{other_paras}{a list of other curvep parameters to pass on}

\item{vehicle_data}{NULL or a dbl vector of responses in vehicle control wells}

\item{simplify_output}{default = FALSE, if set TRUE (withdraw(out, "act") is performed)}
}
\value{
if simplify_output = FALSE, an object of class 'rcurvep_out_nested' (also a tbl)
if simplify_output = TRUE, an object of class 'rcurvep_out' (also a tbl)

\itemize{
  \item input: a tbl, including the information of concs, resps, and parameters
  \item output: a list, all results from `curvep()`
  \item activity: a tbl, activity information from output
  \item endpoint: input endpoint information
  \item chemical: input chemical information
  \item direction: direction used in the calculation
  \item threshold: threshold used in the calculation
  \item repeat_id: repeat id, NA for the calculation using original responses instead of the simulated responses
}
}
\description{
Given a type of dataset and parameters,
the function generates input for `curvep()`,
performs calculations, and generates output.
}
\details{
\code{\link{zfishdev}} for dichotomous binary incidence data and \code{\link{zfishbeh}} for continuous data
}
\examples{

# a binary incidence dataset
data(zfishdev)
x <- zfishdev \%>\% split(.$endpoint)
outd <- run_curvep_batch(x[[1]],
                      directionality = 1,
                      n_sample = 1,
                      threshold = 15,
                      other_paras = list(CARR = 20, TrustHi = TRUE))
# more examples are availabie
vignette("Rcurvep-intro")

}
\seealso{
\code{\link{curvep}} for available curvep parameters

\code{\link{withdraw.rcurvep_out_nested}} if simplify_output = TRUE
}
