% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_onerun.R
\name{run_rcurvep}
\alias{run_rcurvep}
\title{Run rcurvep on a base dataset}
\usage{
run_rcurvep(d, mask = 0, config = curvep_defaults(),
  keep_sets = c("act_set", "resp_set", "fp_set"), ...)
}
\arguments{
\item{d}{a dataset with columns: endpoint, chemical, conc, and resp, mask (optional) (see \code{\link{zfishbeh}})}

\item{mask}{(default = 0, not using mask); if mask column exists, the setting will be ignored
use a vector of integers to mask the resps, 1 to mask resp at the highest conc, 2 to mask resp at the second highest conc, and so on.}

\item{config}{default values are by curvep_defaults()}

\item{keep_sets}{act_set + (resp_set or fp_set) or all three
\itemize{
  \item act_set: activity data
  \item resp_set: response data
  \item fp_set: fingerprint data
}}

\item{...}{a list of settings; can be used to overwrite the default values in curvep_defaults()}
}
\value{
a rcurvep object; it has two components: result, config
}
\description{
Run rcurvep on a base dataset
}
\examples{

data(zfishbeh)
d <- create_dataset(zfishbeh)

# default
out <- run_rcurvep(d)

# change THR
out <- run_rcurvep(d, TRSH = 30)

# mask response at highest and second highest concentration
out <- run_rcurvep(d, mask = c(1, 2))


}
