% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_simulate.R
\name{simulate_resps}
\alias{simulate_resps}
\title{Simulate responses}
\usage{
simulate_resps(d, dataset_type, n_samples, vdata = NULL)
}
\arguments{
\item{d}{a dataset}

\item{dataset_type}{continuous or dichonomous}

\item{n_samples}{number of samples}

\item{vdata}{vehicle contol response data}
}
\value{
d with sample_id and new resp columns
}
\description{
Simulate responses
}
\keyword{internal}
