% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{summarize_actset_by_type}
\alias{summarize_actset_by_type}
\title{Summarize the activity columns based on type}
\usage{
summarize_actset_by_type(act_set_grouped, type, ci_level)
}
\arguments{
\item{act_set_grouped}{a grouped act_set}

\item{type}{either med_only, med_conf, hit_conf}

\item{ci_level}{confidence interval level}
}
\value{
a tibble with added columns (_med, _ciu, _cil, hit_confidence) depending on the type
}
\description{
Summarize the activity columns based on type
}
\keyword{internal}
