% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{summarize_actset_in}
\alias{summarize_actset_in}
\title{Summarize an active_set}
\usage{
summarize_actset_in(act_set, ci_level)
}
\arguments{
\item{act_set}{one active_set}

\item{ci_level}{confidence interval}
}
\value{
a tibble with summarized columns
}
\description{
Summarize an active_set
}
\keyword{internal}
