% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{summarize_actsets}
\alias{summarize_actsets}
\title{Summarize act_sets}
\usage{
summarize_actsets(nestd, ci_level)
}
\arguments{
\item{nestd}{the nested tibble from the get_nested_joined_sets}

\item{ci_level}{confidence level}
}
\value{
the original tibble with a new column act_summary
}
\description{
Summarize act_sets
}
\keyword{internal}
