% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvep_out_extraction.R
\name{summarize_curvep_output}
\alias{summarize_curvep_output}
\title{Summarize the activity output after curvep based on bootstrap samples.
The summarized values include hit_confidence, confidence interval for the activity values,
and median responses per concentration.}
\usage{
summarize_curvep_output(a_out, col_names = c("POD", "EC50", "Emax", "wAUC",
  "wAUC_prev"), modifier = NULL, conf_level = 0.95)
}
\arguments{
\item{a_out}{a tibble from the extract_curvep_data(type != "summary") or run_curvep_job(simplify_output = TRUE)}

\item{col_names}{default values include POD, EC50, Emax, wAUC, wAUC_prev}

\item{modifier}{A string to match the Curvep Comments column to batch modifiy the activity,
including hit, potency related values, and adjusted responses (active -> inactive)}

\item{conf_level}{default value = 0.95 (95\%)}
}
\value{
a tibble with summary data: 1) the hit confidence, 2) the median (med),
  95\% confidence interval (ciu, cil) of activity values, 3) median of responses of concentrations (input/output)
3) is only calculated when input tables have resps, concs, and resps_in columns
}
\description{
Summarize the activity output after curvep based on bootstrap samples.
The summarized values include hit_confidence, confidence interval for the activity values,
and median responses per concentration.
}
\examples{
data(zfishdev)
x <- zfishdev \%>\% split(.$endpoint)
acts <- run_curvep_job(x[[1]],
                      directionality = 1,
                      n_sample = 10,
                      threshold = 15,
                      other_paras = list(CARR = 20, TrustHi = TRUE), simplify_output = TRUE)
# activities
acts_sum <- summarize_curvep_output(acts, conf_level = 0.9)

}
