% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_simulate.R
\name{summarize_resps}
\alias{summarize_resps}
\title{Summarize responses}
\usage{
summarize_resps(d, dataset_type)
}
\arguments{
\item{d}{a dataset such as \code{\link{zfishdev}} and \code{\link{zfishbeh}}}

\item{dataset_type}{continuous or dichotomous}
}
\value{
d with a modified or created resp column
}
\description{
Summarize responses
}
\keyword{internal}
