% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvep_out_extraction.R
\name{summary.rcurvep_out}
\alias{summary.rcurvep_out}
\title{Summarize the activity output after curvep based on simulated samples.}
\usage{
\method{summary}{rcurvep_out}(a_out, col_names = c("POD", "EC50", "Emax",
  "wAUC", "wAUC_prev"), modifier = NULL, conf_level = 0.95)
}
\arguments{
\item{a_out}{a tbl from the `withdraw()`(type != "summary") or `run_curvep_batch(simplify_output = TRUE)`}

\item{col_names}{default activity values include POD, EC50, Emax, wAUC, wAUC_prev}

\item{modifier}{a chr to match the curvep comments column to batch-modify the activity,
including hit, potency related values, and adjusted responses (active -> inactive)}

\item{conf_level}{default value = 0.95 (95\%)}
}
\value{
a tbl with summary data: 1) the hit confidence, 2) the median (med),
  95\% confidence interval (ciu, cil) of activity values, 3) the median response per concentration (input/output)
3) is only calculated when the input tbl have resps, concs, and resps_in columns
}
\description{
The summarized values include hit_confidence, confidence interval of the activity values,
and the median response per concentration.
}
\examples{
data(zfishdev)
x <- zfishdev \%>\% split(.$endpoint)
acts <- run_curvep_batch(x[[1]],
                      directionality = 1,
                      n_sample = 10,
                      threshold = 15,
                      other_paras = list(CARR = 20, TrustHi = TRUE), simplify_output = TRUE)
# activities
acts_sum <- summary(acts, conf_level = 0.9)

}
