% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{unnest_joined_sets}
\alias{unnest_joined_sets}
\title{Unnest the nested tibble with columns act_set, resp_set, fp_set, or act_summary}
\usage{
unnest_joined_sets(nested, base_cols, add_col)
}
\arguments{
\item{nested}{a nested tibble with columns}

\item{base_cols}{a vector of common column names in sets}

\item{add_col}{act_set, resp_set, fp_set, or act_summary}
}
\value{
a tibble
}
\description{
Unnest the nested tibble with columns act_set, resp_set, fp_set, or act_summary
}
\keyword{internal}
