% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{unnest_knee_data}
\alias{unnest_knee_data}
\title{Unnest the knee_out object}
\usage{
unnest_knee_data(kneed, type = c("stats", "outcome"))
}
\arguments{
\item{kneed}{output from cal_dataset_knee()}

\item{type}{either stats or outcome}
}
\value{
a tibble
}
\description{
Unnest the knee_out object
}
\keyword{internal}
