% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvep_out_extraction.R
\name{withdraw}
\alias{withdraw}
\alias{withdraw.rcurvep_out_nested}
\title{Withdraw curvep input, output, or activity data}
\usage{
withdraw(c_out, type, modifier = NULL)

\method{withdraw}{rcurvep_out_nested}(c_out, type, modifier = NULL)
}
\arguments{
\item{c_out}{a tbl after `run_curvep_batch()`.}

\item{type}{a chr to indicate which data to withdraw
\itemize{
  \item act: all the activity related metrics such as potency and efficacy
  \item concs_in: a vector of input concentrations
  \item resps_in: a vector of input resps (simulation/original)
  \item resps_out: a vector of output (clean) responses
  \item paras_in: user manually input parameters
  \item paras: all the parameters used in the calculation
  \item summary: 1) the hit confidence, 2) the median (med),
  95\% confidence interval (ciu, cil) of POD, EC50, Emax, and wAUC, 3) median of resps (input/output)
}}

\item{modifier}{a chr to match the curvep Comments column to batch-modify the activity,
including hit, potency related values, and adjusted responses (active -> inactive)}
}
\value{
Depending the specified type, a tbl with respective columns is returned.
}
\description{
The function unnests the input, output, or activity lists after `run_curvep_batch()`.
}
\examples{
data(zfishdev)
x <- zfishdev \%>\% split(.$endpoint)
outd <- run_curvep_batch(x[[1]],
                      directionality = 1,
                      n_sample = 1,
                      threshold = 15,
                      other_paras = list(CARR = 20, TrustHi = TRUE))
# activities
withdraw(outd, "act")

# parameters
withdraw(outd, "paras")

# summary
withdraw(outd, "summary")

# make comments that contain "INVERSE" string as inactive
withdraw(outd, "act", "INVERSE")



}
