% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zfishbeh}
\alias{zfishbeh}
\title{Subsets of concentration response datasets from zebrafish neurotoxicity assays}
\format{
A tibble with 2123 rows and 4 columns:
\describe{
\item{endpoint}{endpoint name}
\item{chemical}{chemical name + CASRN}
\item{conc}{concentrations in log10(M) format}
\item{resp}{responses after normalized using the vehicle control on each plate}
}
}
\source{
Biobide study S-BBD-0017/15
}
\usage{
zfishbeh
}
\description{
The datasets contain 11 toxicity endpoints and 2 chemicals.
The responses have been normalized so that the baseline is 0.
}
\keyword{datasets}
