% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zfishdev_act}
\alias{zfishdev_act}
\title{Activity output based on simulated datasets using zfishdev_all dataset}
\format{A list of two named components: result and config.
The result component is a list with one component: act_set.}
\usage{
zfishdev_act
}
\description{
The data is an rcurvep object from the \code{\link[=combi_run_rcurvep]{combi_run_rcurvep()}}.
See \code{\link[=combi_run_rcurvep]{combi_run_rcurvep()}} for the code to reproduce this dataset.
}
\seealso{
\code{\link[=estimate_dataset_bmr]{estimate_dataset_bmr()}}
}
\keyword{datasets}
