% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zfishdev_all}
\alias{zfishdev_all}
\title{Full sets of concentration response datasets from zebrafish developmental toxicity assays}
\format{A tibble with 512 rows and 5 columns:}
\source{
Biobide study S-BBD-00016/15
}
\usage{
zfishdev_all
}
\description{
The datasets contain 4 toxicity endpoints and 32 chemicals.
}
\seealso{
\link{zfishdev}
}
\keyword{datasets}
