% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR_calculations.R
\name{cal_knee_point}
\alias{cal_knee_point}
\title{Calculate the knee point on the exponential-like curve using various approaches}
\usage{
cal_knee_point(df, xvar, yvar, p1_raw = NULL, p2_raw = NULL)
}
\arguments{
\item{df}{a tbl}

\item{xvar}{A chr column name in the `df` to be the x-axis in the exponential-like curve}

\item{yvar}{A chr column name in the `df` to be the y-axis in the exponential-like curve}

\item{p1_raw}{default = NULL, or an int to manually set the first threshold index for the original value-based approach}

\item{p2_raw}{default = NULL, or an int to manually set the last threshold index for the original value-based approach}
}
\value{
same as \code{\link{get_baseline_threshold}}
}
\description{
Two approaches are available: original value-based and exponential fit value-based.
Currently original value-based approach appears to be more consistent.
}
\examples{
data("zfishdev_act")
outthres <- get_baseline_threshold(zfishdev_act)[["stats"]]

outthres2 <- outthres \%>\%
  nest(-endpoint, -direction) \%>\%
  mutate(
   knee_out = purrr::map(data, function(x) cal_knee_point(x, "threshold", "pooled_variance"))
  )

}
