% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_bmr.R
\name{cal_knee_point}
\alias{cal_knee_point}
\title{Calculate the knee point on the exponential-like curve using various approaches}
\usage{
cal_knee_point(d, xaxis, yaxis, p1 = NULL, p2 = NULL, plot = TRUE)
}
\arguments{
\item{d}{a tibble}

\item{xaxis}{a chr column name in the `d` to be the x-axis in the exponential-like curve}

\item{yaxis}{a chr column name in the `d` to be the y-axis in the exponential-like curve}

\item{p1}{default = NULL, or an int to manually set the first index of line}

\item{p2}{default = NULL, or an int to manually set the last index of line}
}
\value{
a list with two components: stats and outcome
}
\description{
Two approaches are available: original value-based and exponential fit value-based
}
\examples{

inp <- tibble(
x = seq(5, 95, by = 5),
y = c(0.0537, 0.0281, 0.0119, 0.0109, 0.0062, 0.0043, 0.0043, 0.0042, 0.0041, 0.0043, 0.0044, 0.0044, 0.0046, 0.0051, 0.0055, 0.0057, 0.0072, 0.0068, 0.0035)
)

out <- cal_knee_point(inp,"x", "y", plot = FALSE)
plot(out)


}
