% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvep.R
\name{curvep}
\alias{curvep}
\title{Curvep function}
\usage{
curvep(Conc, Resp, Mask = NULL, TRSH = 15, RNGE = -100, MXDV = 5,
  CARR = 0, BSFT = 3, USHP = 4, TrustHi = FALSE, StrictImp = TRUE,
  DUMV = -999, TLOG = -24)
}
\arguments{
\item{Conc}{array of concentrations, e.g., in Molar units, can be log-transformed, in which case internal log-transformation is skipped}

\item{Resp}{array of responses at corresponding concentrations, e.g., raw measurements or normalized to controls}

\item{Mask}{array of 1/0 flags indicating invalidated measurements (default = NULL)}

\item{TRSH}{base(zero-)line threshold (default = 15)}

\item{RNGE}{target range of responses (default = -100)}

\item{MXDV}{maximum allowed deviation from monotonicity (default = 5)}

\item{CARR}{carryover detection threshold (default = 0, analysis skipped if set to 0)}

\item{BSFT}{for baseline shift issue, min.#points to detect baseline shift (default = 3, analysis skipped if set to 0)}

\item{USHP}{for u-shape curves, min.#points to avoid flattening (default = 4, analysis skipped if set to 0)}

\item{TrustHi}{for equal sets of corrections, trusts those retaining measurements at high concentrations (default = FALSE)}

\item{StrictImp}{prevents extrapolating over concentration-range boundaries; used for POD, ECxx etc (default = TRUE)}
}
\value{
a list with corrected concentration-response measurements and several calculated curve metrics
\itemize{
  \item resp corrected responses
  \item corr flags for corrections
  \item ECxx effective concentration values at various thresholds
  \item Cxx concentrations for various absolute response levels
  \item Emax maximum effective concentration, slope of the mid-curve (b/w EC25 and EC75)
  \item wConc, response-weighted concentration
  \item wResp concentration-weighed response
  \item POD point-of-departure (first concentration with response >TRSH)
  \item AUC area-under-curve (in units of log-concentration X response)
  \item wAUC AUC weighted by concentration range and POD / TLOG (-12)
  \item wAUC_pre AUC weighted by concentration range and POD
  \item nCorrected number of points corrected (basically, sum of flags in corr)
  \item Comments warning and notes about the dose-response curve
  \item Settings input parameters for this run
}
}
\description{
A function to process a given pair of concentration response data, the background function for `run_curvep_job()`.
}
\section{primary parameters}{

}

\section{secondary parameters}{

}

\examples{
curvep(Conc = c(-8, -7, -6, -5, -4) , Resp = c(0, -3, -5, -15, -30))
}
\seealso{
\code{\link{run_curvep_job}}
}
