% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_onerun.R
\name{curvep_defaults}
\alias{curvep_defaults}
\title{Default configuration for curvep}
\usage{
curvep_defaults()
}
\value{
a list of parameters with class as curvep_config

\itemize{
  \item TRSH base(zero-)line threshold
  \item RNGE arget range of responses
  \item MXDV maximum allowed deviation from monotonicity
  \item CARR carryover detection threshold (analysis skipped if set to 0)
  \item BSFT for baseline shift issue, min.#points to detect baseline shift (analysis skipped if set to 0)
  \item USHP for u-shape curves, min.#points to avoid flattening (analysis skipped if set to 0)
  \item TrustHi for equal sets of corrections, trusts those retaining measurements at high concentrations
  \item StrictImp prevents extrapolating over concentration-range boundaries; used for POD, ECxx etc.
  \item DUMV dummy value for inactive (not suggested to modify)
  \item TLOG denominator for calculation wAUC (not suggested to modify)
}
}
\description{
curvep parameters
}
\examples{

# display all default settings
curvep_defaults()

# customize settings
custom_settings <- curvep_defaults()
custom_settings$TRSH <- 30
custom_settings



}
