% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvep_out_extraction.R
\name{extract_curvep_data}
\alias{extract_curvep_data}
\title{Unnest Curvep input, output, or activity lists}
\usage{
extract_curvep_data(c_out, type)
}
\arguments{
\item{c_out}{A tibble after `run_curvep_job()`.}

\item{type}{A string to indicate which data to extract. Currently seven types are implemented:
\itemize{
  \item act: all the activity related metrics such as potency and efficacy
  \item concs_hl: the highest and lowest tested concentration
  \item concs_in: a vector of input concentrations
  \item resps_out: a vector of output (clean) responses
  \item paras_in: user manually input parameters
  \item paras: all the parameters used in the calculation
  \item summary: 1) the hit confidence, 2) the median (med),
  95\% confidence interval (ciu, cil) of POD, EC50, Emax, and wAUC, 3) median of concs/resps,
  based on bootstrap samples
}}
}
\value{
Depending the specified type, a tibble with various columns is returned.
}
\description{
The function unnest the input, output, or activity lists after `run_curvep_job()`
}
\examples{
data(zfishdev)
x <- zfishdev \%>\% split(.$endpoint)
outd <- run_curvep_job(x[[1]],
                      directionality = 1,
                      n_sample = 1,
                      threshold = 15,
                      other_paras = list(CARR = 20, TrustHi = TRUE))
# activities
extract_curvep_data(outd, "act")

# parameters
extract_curvep_data(outd, "paras")

# summary
extract_curvep_data(outd, "summary")

}
\seealso{
\code{\link{curvep}} for available Curvep parameters
}
