% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvep_out_extraction.R
\name{extract_curvep_data}
\alias{extract_curvep_data}
\title{Unnest Curvep input, output, or activity lists}
\usage{
extract_curvep_data(c_out, type, modifier = NULL)
}
\arguments{
\item{c_out}{A tibble after `run_curvep_job()`.}
}
\value{
Depending the specified type, a tibble with various columns is returned.
}
\description{
The function unnest the input, output, or activity lists after `run_curvep_job()`
}
\examples{
data(zfishdev)
x <- zfishdev \%>\% split(.$endpoint)
outd <- run_curvep_job(x[[1]],
                      directionality = 1,
                      n_sample = 1,
                      threshold = 15,
                      other_paras = list(CARR = 20, TrustHi = TRUE))
# activities
extract_curvep_data(outd, "act")

# parameters
extract_curvep_data(outd, "paras")

# summary
extract_curvep_data(outd, "summary")

# make comments that contain "INVERSE" string as inactive

extract_curvep_data(outd, "act", "INVERSE")

}
\seealso{
\code{\link{curvep}} for available Curvep parameters
}
