% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMR_calculations.R
\name{identify_basenoise_threshold}
\alias{identify_basenoise_threshold}
\title{Identify baseline noise threshold through activity data from simulated curves}
\usage{
identify_basenoise_threshold(df, endpoint = "endpoint",
  chemical = "chemical", threshold = "threshold", direction = "direction",
  potency = "POD", plot = TRUE, n_endpoint_page = 4)
}
\arguments{
\item{df}{a tibble}

\item{endpoint}{a string that represents the column name of endpoint in `df`}

\item{chemical}{a string that represents the column name of chemical in `df`}

\item{threshold}{a string that represents the column name of threshold in `df`}

\item{direction}{a string that represents the column name of direction in `df`}

\item{potency}{a string that represents the column name of potency in `df`}

\item{plot}{default = TRUE, diagnostic plot}

\item{n_endpoint_page}{number of endpoints to be plotted per page}
}
\value{
a list with two tibbles

\describe{
  \item{stats}{
  y_exp_fit, y_lm_fit: the fitted y values (exponential or linear)
  dist2l_exp, dist2l_raw: the calculated distance-to-line value based on fitted y values (exponential) or raw y values
  curvature: the calculated curvature value using smooth.spline fit results
  p1_raw, p2_raw: the index of points for the line using the raw y values
  }
  \item{outcome}{
  thresDist_raw: the identified threshold based on distance approach using the raw y values
  thresDist_exp: the identified threshold based on distance approach using the y values from the fitted exponential curve
  thresComment: the flag to suggest whether to use the thresDist (OK, cautionary, check)
  p1_raw, p2_raw: the index of points for the line using the raw y values
  cor_exp_fit, cor_lm_fit: the Pearson's correlation between the fitted y values (exponential or linear) and the raw y values
  }
}
}
\description{
Provided an activity dataset after `run_curvep_job()` and `extract_curvep_data()`,
the function calculates the pooled variance of potency information (i.e., POD parameter) across chemicals
and approaches are used to derive the lowest threshold that variance of POD is sufficiently reduced and even stabilized.
}
\examples{
data("zfishdev_act")
acts <- zfishdev_act \%>\% mutate(POD = ifelse(is.na(POD), conc_highest, POD))
outthres <- identify_basenoise_threshold(acts)

}
