% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_plot.R
\name{plot.rcurvep_bmr}
\alias{plot.rcurvep_bmr}
\title{Plot BMR diagnostic curves}
\usage{
\method{plot}{rcurvep_bmr}(x, n_in_page = 6)
}
\arguments{
\item{x}{the rcurvep_bmr object (see\code{\link{estimate_dataset_bmr}})}

\item{n_in_page}{number of endpoints in a page}
}
\value{
a ggplot object
}
\description{
Plot BMR diagnostic curves
}
\examples{
data(zfishdev_act)

# use the highest concentration to replace value of the inactive
sumd <- summarize_rcurvep_output(zfishdev_act, clean_only = TRUE)

bmr_out <- estimate_dataset_bmr(sumd, plot = FALSE)
plot(bmr_out)

}
