% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcurvep-package.R
\docType{package}
\name{Rcurvep}
\alias{Rcurvep}
\alias{Rcurvep-package}
\title{Rcurvep: a R package for concentration response modeling using Curvep}
\description{
The packages provides an R interface for processing concentration response data using Curvep, a response noise filtering algorithm.
}
\details{
With a predefined response noise threshold in an experiment, it allows to calculate activity based on original or simulated concentration response data.
The simulation process differs for three types of dataset:
\itemize{
  \item binary incidece data (e.g., mortality data)
  \item high number of replicates (e.g., alternative animal model data)
  \item low number of replicates (e.g., in vitro data)
}
If the response noise threshold is unknown, the above process can be repeated using a number of candidate thresholds.
The optimal threshold is identified as the one that varience of potency estimation is sufficiently reduced and even stabilized.

To learn more about Rcurvep, start with the vignettes:
`browseVignettes(package = "Rcurvep")`
}
