% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvep_job_wrap.R
\name{run_curvep_job}
\alias{run_curvep_job}
\title{A wrap function to run Curvep based on types of dataset}
\usage{
run_curvep_job(dats, directionality = c(1, 0, -1), n_sample = NULL,
  threshold, other_paras = list(), vehicle_data = NULL,
  simplify_output = FALSE)
}
\arguments{
\item{dats}{datasets such as \code{\link{zfishdev}} and \code{\link{zfishbeh}}}

\item{directionality}{an int value (1, 0, -1) to represent the presumed direction of responses for processing; 1 = up, -1 = down, 0 = both}

\item{n_sample}{NULL (using orignal data) or an int to indicate the number of curves to generate}

\item{threshold}{a numeric value for the presumed noise threshold or a numeric vector for threshold finding or a named (-1, 1) list with numeric vectors}

\item{other_paras}{a list of other Curvep parameters to pass on}

\item{vehicle_data}{NULL or a numeric vector of responses in vehicle control wells}

\item{simplify_output}{(default = FALSE) warning! large output, set TRUE (extract_curvep_data(out, "act") is performed) for threshold finding calculations}
}
\value{
\itemize{
  \item input: a tibble, including the information of concs, resps, and parameters
  \item output: a list, all results from `curvep()`
  \item activity: a tibble, extracted activity information from output
  \item endpoint: input endpoint information
  \item chemical: input chemical information
  \item direction: direction used in the calculation
  \item threshold: threshold used in the calculation
  \item repeat_id: repeat id, NA for the calculation using orignal responses instead of simulated samples
}
}
\description{
Given a type of dataset and parameters,
the function generates input for `curvep()`,
performs calculations, and generates output.
}
\examples{

# a percentage type dataset
data(zfishdev)
x <- zfishdev \%>\% split(.$endpoint)
outd <- run_curvep_job(x[[1]],
                      directionality = 1,
                      n_sample = 1,
                      threshold = 15,
                      other_paras = list(CARR = 20, TrustHi = TRUE))
# more examples are availabie
vignette("Rcurvep-intro")

}
\seealso{
\code{\link{curvep}} for available Curvep parameters
}
