% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvep_job_wrap.R
\name{run_curvep_job}
\alias{run_curvep_job}
\title{A wrap function to run Curvep based on types of datasets}
\usage{
run_curvep_job(dats, directionality = c(1, 0, -1), n_sample, threshold,
  other_paras = list(), vehicle_data = NULL)
}
\arguments{
\item{dats}{datasets}

\item{threshold}{a numeric value for the presumed noise threshold or a numeric vector for threshold finding}

\item{other_paras}{(optional) a list of other Curvep parameters to pass on}

\item{vehicle_data}{NULL or a numeric vector of responses in vehicle control wells}
}
\value{
\itemize{
  \item dduid: dataset unique ID, it is the combination of endpoint, chemical, and directionality (from input) as well as directionality used (from calculation).
  \item input: a tibble, including the information of concs, resps, and parameters
  \item output: a list, all results from `curvep()`
  \item activity: a tibble, extracted activity information from output
  \item repeat_id: repeat id, NA for the calculation using orignal responses instead of bootstrap samples.
  \item thres: threshold used in the calculation
}
}
\description{
Given a type of datasets and parameters,
the function generates input for `curvep()`,
performs calculations, generates output, and extracts activity.
}
\examples{

# a percentage type dataset
data(zfishdev)
x <- zfishdev \%>\% split(.$endpoint)
outd <- run_curvep_job(x[[1]],
                      directionality = 1,
                      n_sample = 1,
                      threshold = 15,
                      other_paras = list(CARR = 20, TrustHi = TRUE))
# more examples are availabie `browseVignettes(package = "Rcurvep")`

}
\seealso{
\code{\link{zfishdev}}

\code{\link{zfishbeh}}

\code{\link{curvep}} for available Curvep parameters
}
