% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurvep_summary.R
\name{summarize_rcurvep_output}
\alias{summarize_rcurvep_output}
\title{Clean and summarize the output of rcurvep}
\usage{
summarize_rcurvep_output(d, act_modifier = NULL, ci_level = 0.95,
  clean_only = FALSE)
}
\arguments{
\item{d}{the list from the \code{\link{combi_run_rcurvep}} or \code{\link{run_rcurvep}}}

\item{act_modifier}{a character string, default = NULL,
to make the curve with this string in the Comments column as inactive.
The most common case to unhit is the "INVERSE" curves.}

\item{ci_level}{default = 0.95 (95% confidence interval)}

\item{clean_only}{default = FALSE, only the 1st, 2nd, 3rd task will be performed}
}
\value{
the original list (with modified components) + act_summary
}
\description{
1. adds an hit column in the act_set
2. unhit (make result as inactive) if the Comments column contains a certain string
3. make NA in the potency related columns as highest tested concentration
4. summarize the results based on the ci_level
}
\examples{

data(zfishbeh)

# base dataset
out <- combi_run_rcurvep(zfishbeh, n_samples = NULL, TRSH = c(5, 10))
out_res <- summarize_rcurvep_output(out)

# base dataset + remove comment has "INVERSE"
out <- summarize_rcurvep_output(out, act_modifier = "INVERSE")

# simulated dataset
out <- combi_run_rcurvep(zfishbeh, n_samples = 3, TRSH = c(5, 10))
out_res <- summarize_rcurvep_output(out)

}
