% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zfishbeh}
\alias{zfishbeh}
\title{Zebrafish behavior endpoint data}
\format{a tibble with variables
\itemize{
\item endpoint toxicity endpoints
\item chemical chemical information
\item concs concentrations in log10(M) format
\item resps normalized responses with baseline = 0
}}
\source{
Biobide study S‐BBD‐0017/15
}
\usage{
zfishbeh
}
\description{
Data are processed from Biobide behavior plate data
}
\keyword{datasets}
