% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zfishdev}
\alias{zfishdev}
\title{Zebrafish development endpoint data}
\format{a tbl with columns:
\itemize{
\item endpoint: toxicity endpoints
\item chemical: chemical information
\item concs: concentrations in log10(M) format
\item n_in: number of incidence
\item N: number of animals
}}
\source{
Biobide study S-BBD-00016/15
}
\usage{
zfishdev
}
\description{
Data are processed from Biobide development plate incidence data
}
\keyword{datasets}
