% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zfishdev_act}
\alias{zfishdev_act}
\title{Zebrafish development endpoint activity results}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1216000 rows and 20 columns.}
\usage{
zfishdev_act
}
\description{
zfishdev_complete data are processed through `run_curvep_job()` and `extract_curvep_data()`
}
\keyword{datasets}
