% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zfishdev_act}
\alias{zfishdev_act}
\title{Zebrafish development endpoint activity results}
\format{An object of class \code{rcurvep_out} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 121600 rows and 20 columns.}
\usage{
zfishdev_act
}
\description{
zfishdev_act data are processed through `run_curvep_batch()` and `withdraw()`
}
\keyword{datasets}
