% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zfishdev_all}
\alias{zfishdev_all}
\title{Zebrafish development data for all chemicals}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 512 rows and 5 columns.}
\usage{
zfishdev_all
}
\description{
Zebrafish development data for all chemicals
}
\keyword{datasets}
